Fixing Client side JS
Incorportaed changes to modal by Dakota.
I also add some files to manage all the JS code so we don't end up with
code that doesn't apply to a page. I set up the templates so that if you
pass the template a user ID it will display it and logoff but it
displays logon. also set up index.js for loggin in and logging off and
creating a new user i am not currently sending any thing to the server
beacuse I'm not sure how the logon process looks yet but when we know i
will do it.